+++ validationframework-core/src/main/java/com/google/code/validationframework/base/common/CompositeThrowableHandler.java
@@ -25,42 +25,18 @@
 
 package com.google.code.validationframework.base.common;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 /**
- * Unchecked exception handler logging an warning message to handled the runtime exceptions and errors.
+ * Composite throwable handler.
  *
- * @see UncheckedExceptionHandler
+ * @param <T> Type of throwable that can be handled by all handlers.
  */
-public class LogWarningUncheckedExceptionHandler implements ThrowableHandler<Throwable>, UncheckedExceptionHandler {
-
-    /**
-     * Logger for this class.
-     */
-    private static final Logger LOGGER = LoggerFactory.getLogger(LogWarningUncheckedExceptionHandler.class);
-
-    /**
-     * @see UncheckedExceptionHandler#handleException(Exception)
-     */
-    @Override
-    public void handleException(RuntimeException exception) {
-        handleThrowable(exception);
-    }
-
-    /**
-     * @see UncheckedExceptionHandler#handleError(Error)
-     */
-    @Override
-    public void handleError(Error error) {
-        handleThrowable(error);
-    }
+public class CompositeThrowableHandler<T extends Throwable> implements ThrowableHandler<T> {
 
     /**
      * @see ThrowableHandler#handleThrowable(Throwable)
      */
     @Override
-    public void handleThrowable(Throwable throwable) {
-        LOGGER.warn("An exception or error occurred", throwable);
+    public void handleThrowable(T throwable) {
+
     }
 }