+++ validationframework-core/src/main/java/com/google/code/validationframework/base/common/UncheckedExceptionHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015, ValidationFramework Authors
+ * Copyright (c) 2016, ValidationFramework Authors
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -29,9 +29,12 @@ package com.google.code.validationframework.base.common;
  * Exception handler to process runtime exceptions and errors.
  *
  * @see ExceptionHandler
+ * @see ThrowableHandler
  * @see RuntimeException
  * @see Error
+ * @deprecated For simplification, you may use the {@link ThrowableHandler} instead.
  */
+@Deprecated
 public interface UncheckedExceptionHandler extends ExceptionHandler<RuntimeException> {
 
     /**