+++ CHANGES.md
@@ -22,6 +22,7 @@
 * Deprecated the experimental `ResultCollectorValidatorBuilder` in favor of the `GeneralValidatorBuilder`
 * Removed experimental duplicate of `PropertyResultHandler`
 * Removed experimental duplicate of `JFormattedTextFieldValueProperty `
+* Deprecated the ExceptionHandler and the UncheckedExceptionHandler in favor of the new ThrowableHandler
 * Added `PseudoClassResultHandler` to JavaFX support
 
 ## Backward incompatibilities in terms of compilation